slid 18 Erorr DY2 Slide 25; Error on saving/viewing in data tables DY3 Slide 30; Fixed viewing/ removed any published/ishing DY5 Slide 43: Focused on editing/delte; now comments DY6 Slide 49: Continue on comments DY7 Slide 61: Continue on comments/ works fully and now adding extras DY8 Slide 71: Adding replies/ form on pug might have error DY9 PT2 Slide 7: Doing user_roles DY10 PT2 Slide 19: New Model/ Didn't finish DY11 PT2 Slide 37: Update Controller/ Broken(passport) DY12 PT2 Slide 57: Login/logout function/more middleware function/ prevent admin in works DY13 PT2 Slide 70: Delete Order/ Now Delete Replies

+++ controllers/menuController.js
@@ -51,6 +51,15 @@ module.exports.displayAll = async function(req, res){
     res.render('menu/viewAll', {menu});
 }
 
+module.exports.renderEditForm = async function(req, res){
+    const menu = await Menu.findByPk(req.params.menuId);
+    if (!menu.isOwnedBy(user)){
+        res.redirect('/');
+        return;
+    }
+    res.render('menus/edit', {menu});
+};
+
 module.exports.renderAddForm = async function(req, res){
     const menu = {
         title: '',
@@ -63,6 +72,11 @@ module.exports.renderAddForm = async function(req, res){
 }
 
 module.exports.updateMenu = async function(req, res){
+    const menu = await Menu.findByPk(req.parans.menuId);
+    if (!menu.isOwnedBy(user)){
+        res.redirect('/');
+        return;
+    }
     await Menu.update({
         title: req.body.title,
         intro: req.body.intro,
@@ -78,6 +92,11 @@ module.exports.updateMenu = async function(req, res){
 }
 
 module.exports.deleteMenu = async function(req, res){
+    const menu = await Menu.findByPk(req.params.menuId);
+    if (!user.is('admin') && !menu.isOwnedBy(user)){
+        res.redirect('/');
+        return;
+    }
     await Menu.destroy({
         where: {
             id: req.params.menuId
